Timing update
Per suggestion in documentation it gets the min time not the average. it runs for 1000 times and get the minimum. It is listed in ms.  It uses timeout.default_timer not timeout.repeat bc timeout.repeat would not allow it to return the ciphertext back and printing in the ecryption function took up too much space. Maye we should consider reducing the amount of times it runs as 1000 may be excessive?