encrypts special characters
encrypts printable ascii characters. in the ascii table of printable characters there are 4 sections that have special characters and this encrypts based on the ascii value of the character for special characters.

Note: I adjusted the modulo number for the digits and special characters and it now will only turn a number into a number and likewise for the special characters.